+++ WebSocketServer/.gitignore
@@ -1,1 +1,2 @@
 .env
+build/
