--- WebSocketServer/infrustructure/auth_service/auth_service_init.go
@@ -1,10 +1,0 @@
-package auth_service
-
-import (
-	"WebSocketServer/application/interfaces/auth_service"
-	"WebSocketServer/config"
-)
-
-func GetAuthService(environment *config.Environment) auth_service.AuthServiceInterface {
-	return NewAuthService(environment.JwtSecret)
-}
