+++ WebSocketServer/internal/api/utils/response_marshaling.go
