+++ WebSocketServer/internal/application/dtos/api_dto.go
