+++ WebSocketServer/internal/application/dtos/auth_dto.go
