+++ WebSocketServer/internal/application/dtos/friend_dto.go
