+++ WebSocketServer/internal/application/dtos/tracker_dto.go
