+++ WebSocketServer/internal/config/README.md
