+++ WebSocketServer/internal/infrastructure/auth_service.go
@@ -1,19 +1,19 @@
-package auth_service
+package infrastructure
 
 import (
-	"WebSocketServer/application/dtos"
-	"WebSocketServer/application/interfaces/auth_service"
+	"WebSocketServer/internal/application/dtos"
+	"WebSocketServer/internal/application/interfaces"
 
 	"github.com/golang-jwt/jwt"
 )
 
 type AuthService struct {
-	secret        string
+	secret string
 }
 
 func (service *AuthService) VerifyUser(token string) (*dtos.AuthDetailDTO, error) {
 	claims := jwt.MapClaims{}
-	tkn, err := jwt.ParseWithClaims(token, claims, func(token *jwt.Token)(interface{}, error){
+	tkn, err := jwt.ParseWithClaims(token, claims, func(token *jwt.Token) (interface{}, error) {
 		return []byte(service.secret), nil
 	})
 	if err != nil {
@@ -23,7 +23,7 @@ func (service *AuthService) VerifyUser(token string) (*dtos.AuthDetailDTO, error
 		return nil, err
 	}
 	authDetail := &dtos.AuthDetailDTO{
-		ID: claims["sub"].(string),
+		ID:   claims["sub"].(string),
 		Role: claims["role"].(string),
 		//TODO: Add username to the claims
 		// Username: claims["username"].(string),
@@ -33,8 +33,7 @@ func (service *AuthService) VerifyUser(token string) (*dtos.AuthDetailDTO, error
 
 }
 
-
-func NewAuthService(secret string) auth_service.AuthServiceInterface {
+func NewAuthService(secret string) interfaces.AuthServiceInterface {
 	return &AuthService{
 		secret: secret,
 	}
